% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles-package.R
\docType{package}
\name{smiles-package}
\alias{smiles-package}
\title{smiles: Sequential Method In Leading Evidence Synthesis}
\description{
The purpose of R-package \emph{smiles} is to provide a useful collection of
functions designed to apply sequential method in data synthesis and evidence
evaluation. This package is developed based on \href{https://cran.r-project.org/package=meta}{\emph{meta} (version 7.0-0)} in \bold{R version 4.2.2 (2022-10-31 ucrt)}.
}
\details{
Current version consists of three functions, including a main function of
sequential analysis (i.e. \code{\link{DoTSA}}) and a visualization function
of sequential analysis, (i.e. \code{\link{PlotCoRNET}}).
}
\keyword{package}
